create shared feature extraction functions
